+++ packages/cloud_firestore_odm/cloud_firestore_odm/test/id_test.dart
@@ -1,0 +1,9 @@
+import 'package:cloud_firestore_odm/cloud_firestore_odm.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('Id', () {
+    // ignore: prefer_const_constructors, coverage for the const constructor
+    Id();
+  });
+}