+++ packages/firebase_crashlytics/firebase_crashlytics_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.2.17
+
+ - Update a dependency to the latest release.
+
 ## 3.2.16
 
  - Update a dependency to the latest release.